Got logging working;; Added frontend logic to only do that first write to log if the log file does not already exist;; Working on another backend function to grab the webm audio streams and sort them greatest/least by bit-rate. The next goal will be to display the stream information to each menu, in greatest/least order from left to right window or rather, left being  greatest ALSO planning on using a function-to-run-multiple-functions so I only have to specify a single 'command' in the front end capture_button parameters;; Going to sleeeep now

+++ frontEnd.py
@@ -2,8 +2,9 @@
 
 import os
 import tkinter as tk
-from backEnd import is_valid_url
 import logging
+from tkinter import scrolledtext
+from backEnd import is_valid_url
 from tkinter import messagebox
 
 # Configure Logging
@@ -12,8 +13,9 @@ logging_file = 'ytAudio.log'
 logs = os.path.join(logging_directory, logging_file)
 os.makedirs(logging_directory, exist_ok=True)
 
-with open(logs, 'w') as file:
-    file.write("Log file created")
+if not os.path.exists(logs):
+    with open(logs, 'w') as file:
+        file.write("Log file created")
 
 logging.basicConfig(filename=logs, level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
@@ -45,26 +47,29 @@ url_label.grid(row=1, column=0, pady=10, padx=20, sticky=tk.W)
 url_entry = tk.Entry(window, width=30)
 url_entry.grid(row=1, column=1, pady=10, padx=20, sticky=tk.EW)
 
-stream_menu_high = tk.Frame(window, width=150, height=200, bg="lightblue")
-stream_menu_high.grid(row=2, column=0, pady=20, padx=20, sticky=tk.W)
+##### Stream Menus
+high_menu = tk.Frame(window, width=150, height=200, bg="lightblue")
+high_menu.grid(row=2, column=0, pady=20, padx=20, sticky=tk.W)
 
-stream_menu_medium = tk.Frame(window, width=150, height=200, bg="green")
-stream_menu_medium.grid(row=2, column=1, pady=20, padx=20)
+medium_menu = tk.Frame(window, width=150, height=200, bg="green")
+medium_menu.grid(row=2, column=1, pady=20, padx=20)
 
-stream_menu_low = tk.Frame(window, width=150, height=200, bg="brown")
-stream_menu_low.grid(row=2, column=2, pady=20, padx=20)
+low_menu = tk.Frame(window, width=150, height=200, bg="brown")
+low_menu.grid(row=2, column=2, pady=20, padx=20)
 
 def process_url():
-    entered_url = url_entry.get()
+    url = url_entry.get()
     # check if the entered value is a valid url
-    if is_valid_url(entered_url):
-        logging.info("Success!", "You did it buddy...")
+    if is_valid_url(url):
+        logging.info("The entered URL is a valid, youtube url")
     else:
-        logging.error("Invalid URL...", "Please enter a valid youtube URL...")
+        logging.error("Invalid URL... Please enter a valid youtube URL...")
         messagebox.showerror("Invalid URL...", "Please enter a valid youtube URL...")
 
+# def function_to_run_multiple_functions for my capture button
+
 #### Capture button
-caputure_button = tk.Button(window, text="Process", command=process_url)
-caputure_button.grid(row=1, column=2, pady=10, padx=20, sticky=tk.EW)
+capture_button = tk.Button(window, text="Process", command=process_url)
+capture_button.grid(row=1, column=2, pady=10, padx=20, sticky=tk.EW)
 
 window.mainloop()
